[fix] Fix the benchmark result format
