+++ pyproject.toml
@@ -7,7 +7,7 @@ requires-python = ">=3.9"
 classifiers = [ "Programming Language :: Python :: 3" ]
 readme = {file="README.md", content-type="text/markdown"}
 
-dependencies = [ "resotolib", "resotocore", "resotoworker", "fixcloudutils" ]
+dependencies = [ "resotolib", "resotocore", "resotoworker", "fixcloudutils[arango, redis]" ]
 
 [project.optional-dependencies]
